refacto TestApp to be reusable
