+++ test/layer/polygon_layer_test.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import '../test_utils/test_app.dart';
+
+void main() {
+  testWidgets('test polygon key', (tester) async {
+    await tester.pumpWidget(const TestApp());
+  });
+}
